In AdminController working method delete.

+++ controller/src/main/java/by/it/academy/controller/addon/AdminDelete.java
@@ -12,17 +12,25 @@ import java.io.IOException;
 
 @WebServlet(name = "AdminDelete", value = "/adminDelete")
 public class AdminDelete extends HttpServlet {
+
     private final AdminService adminService = ServiceProvider.getInstance().getAdminService();
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-            Integer idAdminForDelete = Integer.valueOf(request.getParameter("idAdmin"));
-            adminService.deleteAdmin(idAdminForDelete);
-            response.sendRedirect(request.getContextPath() + "/home?pageName=admin");
+        request.getRequestDispatcher("WEB-INF/view/pages/adminDelete.jsp")
+                .forward(request, response);
     }
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
+        if (request.getParameter("cancel") != null) {
+            request.getRequestDispatcher("/home?pageName=admin").forward(request, response);
+        } else {
+            Integer idAdminForDelete = Integer.valueOf(request.getParameter("deleteAdmin"));
+            adminService.deleteAdmin(idAdminForDelete);
+            request.getRequestDispatcher("/home?pageName=admin").forward(request, response);
+
+        }
     }
 }
